requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
dj-database-url==0.5.0
django-bootstrap==0.2.4
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-crispy-forms==1.11.2
django-heroku==0.3.1
gunicorn==20.1.0
install==1.3.4
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.2.0
